requirements.txt
azure-storage-blob
streamlit
python-dotenv
psycopg2-binary
beautifulsoup4
tqdm
google-generativeai
google-ai-generativelanguage
black
isort